Phase 6 -  Observability & Logging - Jenkins Build Logs to Azure Blob/requirements.txt
requests==2.31.0
azure-storage-blob==12.19.0
python-dateutil==2.8.2
